Look for SVGs within engine asset paths
In some situations within Rails 5.1, SVGs cannot be found when they are
coming from within an engine and bundled into the Rails app. It's necessary
to run `rails tmp:cache:clear assets:clobber assets:precompile` a second time
in order to get them to resolve. If the initial compiled asset or
manifest asset can not be found, we'll fall back to looking for the
actual file bundled within a gem somewhere.